docs: :memo: Add docs to Gamepad and Transformations
